going on nbr simpliest way

+++ libftprintf/prs_struct.h
@@ -6,7 +6,7 @@
 /*   By: abarthel <abarthel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/18 16:10:32 by abarthel          #+#    #+#             */
-/*   Updated: 2019/03/13 17:37:45 by abarthel         ###   ########.fr       */
+/*   Updated: 2019/04/01 23:04:33 by abarthel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,9 +18,9 @@
 typedef struct	s_ret
 {
 	char	*ret;
-	long	i;
-	long	fmt_i;
-	long	max;
+	int		i;
+	int		fmt_i;
+	int		max;
 }				t_ret;
 
 extern t_ret	g_ret;
